Multi-platform fixes from yuenicholas-dev.
AlembicMaya.cmake:
Need to change the /MT back to /MD for it to build successfully on Windows with
Visual Studio 2010

Util/Foundation.h:
Need to define _WINSOCKAPI_ so that the correct version of winsock is used for
Windows build

FindIlmBase.cmake:
Add conditional so that we can deal with ILMBase 1.0.3 (which has IexMath) and
previous version which does not.
Add IexMath library to collection.


Fix various Python path issue on Darwin/OSX (support python 2.7 if apple)

Fix typo in referencing Boost's python library, instead of BOOST_PYTHON_LIBRARY,
it should be Boost_PYTHON_LIBRARY