Tweak the IObject class so it can deal with proxies more efficiently.
m_isProxy is not really needed since we need to keep track of the full resolved
path m_proxyFullName.

The recursive walking code has been overhauled to avoid creating vectors which
we constantly copy and mutate.  Instead we can just deal with the target path
and a position that we are searching within it.